requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==3.6.4.0
celery==5.0.0
certifi==2024.2.2
channels==4.1.0
click==8.1.7
click-didyoumean==0.3.1
click-repl==0.3.0
Django==3.2
django-adminlte-3==0.1.6
djangorestframework==3.15.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
frozenlist==1.4.1
h11==0.14.0
idna==3.7
inflection==0.5.1
kombu==5.3.7
magic-filter==1.0.12
multidict==6.0.5
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==24.0
prompt-toolkit==3.0.43
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
yarl==1.9.4
